requirements.txt
aniso8601==10.0.1
attrs==25.3.0
bcrypt==5.0.0
blinker==1.9.0
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-PyMongo==3.0.1
flask-restx==1.3.2
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.2
pymongo==4.15.1
python-dotenv==1.1.1
referencing==0.36.2
rpds-py==0.27.1
typing_extensions==4.15.0
Werkzeug==3.1.3
